backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
psycopg2-binary==2.9.9
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
marshmallow==3.20.2
gunicorn==21.2.0
